+++ src/Apps/AmplSolver/ampl_ipopt.cpp
@@ -83,7 +83,7 @@ int main(
    suffix_handler->AddAvailableSuffix("ipopt_zU_in", AmplSuffixHandler::Variable_Source,
                                       AmplSuffixHandler::Number_Type);
 
-   SmartPtr<TNLP> ampl_tnlp = new AmplTNLP(ConstPtr(app->Jnlst()), app->RegOptions(), app->Options(), args, suffix_handler);
+   SmartPtr<TNLP> ampl_tnlp = new AmplTNLP(ConstPtr(app->Jnlst()), app->RegOptions(), app->Options(), args, suffix_handler, false, NULL, NULL, NULL, NULL, NULL, true);
 
    // Call Initialize again to process output related options
    retval = app->Initialize();